Add option to build only the internal components

+++ setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-usage() { echo "Usage: $0 [-h] [-s] [-p]" 1>&2; exit 1; }
+usage() { echo "Usage: $0 [-h] [-s] [-p] [-i]" 1>&2; exit 1; }
 
 while getopts "hsp" opt; do
     case "${opt}" in
@@ -9,6 +9,9 @@ while getopts "hsp" opt; do
             ;;
         p)
             preserve=1
+            ;;
+        i)
+            internal=1
             ;;
         h|*)
             usage
@@ -25,20 +28,22 @@ if [ -z "${shallow}" ]; then
 
     parent=$(echo $build | sed s/[^\/]*/../g)
 
-    cd deps/zlib
-    rm -rf "$build"
-    mkdir -p "$build" && cd "$build"
-    cmake -G Ninja "$parent"
-    ninja
-    cd "$parent/../../"
+    if [ -z "${internal}" ]; then
+        cd deps/zlib
+        rm -rf "$build"
+        mkdir -p "$build" && cd "$build"
+        cmake -G Ninja "$parent"
+        ninja
+        cd "$parent/../../"
 
-    cd deps/libpng
-    cp scripts/pnglibconf.h.prebuilt pnglibconf.h
-    rm -rf "$build"
-    mkdir -p "$build" && cd "$build"
-    cmake -DPNG_SHARED=OFF -DPNG_TESTS=OFF -DPNG_BUILD_ZLIB=ON -DZLIB_INCLUDE_DIR="$parent/zlib" -G Ninja "$parent"
-    ninja
-    cd "$parent/../../"
+        cd deps/libpng
+        cp scripts/pnglibconf.h.prebuilt pnglibconf.h
+        rm -rf "$build"
+        mkdir -p "$build" && cd "$build"
+        cmake -DPNG_SHARED=OFF -DPNG_TESTS=OFF -DPNG_BUILD_ZLIB=ON -DZLIB_INCLUDE_DIR="$parent/zlib" -G Ninja "$parent"
+        ninja
+        cd "$parent/../../"
+    fi
 fi
 
 rm -rf "$build"
